# api.yaml
openapi: 3.1.0
...
info:
  title: Chatbot Administrasi Mahasiswa
  version: 0.1.0
  description: |
    API chatbot RAG untuk pertanyaan administrasi akademik mahasiswa.

    ## Lima jenis balasan

    Endpoint chat selalu mengembalikan 200 dengan `kind` yang menentukan cara
    frontend menampilkannya. `answer`, `refusal`, dan `support` WAJIB terlihat
    berbeda satu sama lain:

    | `kind` | Arti | Tampilan |
    |---|---|---|
    | `answer` | Jawaban bersumber dokumen resmi | Bubble normal + kartu sitasi (FE-2) |
    | `refusal` | Retrieval terlalu lemah, LLM tidak dipanggil (FR-3) | Visual berbeda + kontak (FE-4) |
    | `support` | Pertanyaan bernuansa tekanan mental (FR-7) | Nada empatik + kontak konseling |
    | `smalltalk` | Sapaan atau basa-basi; retrieval dan LLM dilewati | Bubble biasa, tanpa sitasi dan tanpa tombol umpan balik |
    | `rejected` | Dihentikan gerbang JEV: nonsense, upaya manipulasi, atau di luar topik kampus | Bubble biasa, tanpa sitasi dan tanpa tombol umpan balik |

    Hanya `answer` yang membawa sitasi. Menampilkan kartu sitasi di layar
    penolakan memberi kesan keliru bahwa jawabannya bersumber; menampilkannya
    pada sapaan sama saja mengklaim "halo" ada di panduan akademik.

    `smalltalk` ada supaya mahasiswa yang mengetik "hai" tidak dibalas
    penolakan berisi alamat loket biro, dan supaya sapaan tidak masuk daftar
    pertanyaan tak terjawab (AD-4).

    `rejected` hanya muncul bila gerbang JEV dinyalakan (`JEV_ENABLED=true`).
...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/embed/keys/{key}:
    get:
      tags: [embed]
      summary: Situs yang boleh memuat panel untuk satu kunci sematan
      operationId: get_embed_key
      x-status: implemented
      description: |
        Dipanggil SERVER portal (`client/src/proxy.ts`) setiap kali halaman
        `/embed?key=...` dimuat, bukan oleh peramban. `allowed_origins` menjadi
        header `Content-Security-Policy: frame-ancestors` halaman itu, sehingga
        peramban sendiri yang menolak menampilkan panel di situs lain. Daftar
        kosong = situs mana pun.

        Perubahan dari dashboard berlaku saat panel dibuka berikutnya, tanpa
        build ulang portal.

        Publik: kunci tertulis di kode sumber situs penyemat, dan daftar situsnya
        dapat dilihat siapa pun yang membuka situs-situs itu. Tidak tunduk pada
        kill switch -- panelnya tetap dimuat dan menyampaikan pesan penutupan
        saat mahasiswa bertanya.
      parameters:
        - name: key
          in: path
          required: true
          description: '`emb_` + 24 karakter alfanumerik, dari atribut `data-key` di `embed.js`.'
          schema: { type: string }
      responses:
        '200':
          description: Kunci aktif
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmbedKeyInfo' }
              example:
                allowed_origins: [https://pmb.instiki.ac.id, https://www.instiki.ac.id]
        '404':
          description: Kunci tidak dikenal, salah bentuk, atau sudah dinonaktifkan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/admin/faq:
    get:
      tags: [admin-faq]
      summary: Daftar entri tanya jawab
      operationId: list_faq
      x-status: implemented
      x-min-role: staf
      security: [{ bearerAuth: [] }]
      description: |
        Entri tanya jawab adalah satu pasang pertanyaan dan jawaban yang diketik
        admin langsung di dashboard, tanpa berkas PDF sama sekali.

        Chatbot memperlakukannya persis seperti dokumen: ikut terambil retrieval
        hibrida, ikut berhenti dipakai saat `valid_until` lewat (FR-2), dan ikut
        menjadi kartu sitasi -- dengan `type: tanya_jawab`, sehingga frontend
        tahu tidak ada berkas yang bisa dibuka.

        Inilah jalan tercepat menutup pertanyaan tak terjawab (AD-4): tidak
        semua jawaban perlu menunggu dokumen resmi terbit lebih dulu.

        Staf/dosen hanya melihat entri unitnya, sama seperti dokumen, dan
        `total` pun dihitung atas unit itu.
      parameters:
        - name: include_inactive
          in: query
          schema: { type: boolean, default: false }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Daftar entri tanya jawab
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FaqPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [admin-faq]
      summary: Tambah entri tanya jawab
      operationId: create_faq
      x-status: implemented
      x-min-role: staf
      security: [{ bearerAuth: [] }]
      description: |
        Sinkron seperti unggah dokumen: entri disimpan, dipecah menjadi chunk,
        dan di-embed dalam satu transaksi. Begitu balasan 201 diterima, chatbot
        sudah dapat memakainya pada pertanyaan berikutnya.

        Tulis pertanyaannya seperti mahasiswa akan menanyakannya -- kalimat itu
        ikut diindeks dan ikut tampil sebagai judul sumber pada kartu sitasi.

        Jawaban panjang dipecah menjadi beberapa chunk, dan setiap potongannya
        tetap membawa pertanyaannya.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FaqEntryCreate' }
      responses:
        '201':
          description: Entri tersimpan dan terindeks
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FaqEntry' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '502':
          description: |
            Layanan AI untuk menghitung embedding gagal. `detail` berisi kalimat
            siap tampil yang menyarankan mencoba lagi; entri tidak tersimpan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/admin/faq/{entry_id}:
    parameters:
      - name: entry_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [admin-faq]
      summary: Perbarui entri tanya jawab
      operationId: update_faq
      x-status: implemented
      x-min-role: staf
      security: [{ bearerAuth: [] }]
      description: |
        Mengubah `question` atau `answer` membuang chunk lama dan menghitung
        ulang embedding-nya dalam transaksi yang sama. Indeks yang masih memuat
        kalimat versi lama akan menjawab mahasiswa dengan aturan yang sudah
        dicabut -- persis risiko yang diangkat PRD §12.

        `updated_at` diperbarui saat isi berubah (`question`, `answer`,
        `unit`, `valid_until`), tetapi tidak saat hanya `is_active` yang
        berubah -- sama seperti dokumen.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FaqEntryUpdate' }
      responses:
        '200':
          description: Tersimpan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FaqEntry' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '502':
          description: Layanan AI gagal saat mengindeks ulang; entri tidak berubah.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    delete:
      tags: [admin-faq]
      summary: Hapus entri tanya jawab
      operationId: delete_faq
      x-status: implemented
      x-min-role: staf
      security: [{ bearerAuth: [] }]
      description: |
        Permanen dan tidak dapat dibatalkan; frontend WAJIB meminta konfirmasi
        (PRD §9). Untuk sekadar menghentikan pemakaiannya, pakai
        `PATCH { "is_active": false }`.
      responses:
        '204':
          description: Terhapus
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/admin/embed-keys:
    get:
      tags: [admin-embed-keys]
      summary: Daftar kunci sematan
      operationId: list_embed_keys
      x-status: implemented
      x-min-role: superadmin
      security: [{ bearerAuth: [] }]
      description: |
        Semua kunci, termasuk yang nonaktif, urut dari yang paling lama dibuat,
        beserta jumlah pertanyaan 30 hari terakhir dan waktu pertanyaan terakhir
        dari situs itu.
      responses:
        '200':
          description: Semua kunci
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EmbedKey' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [admin-embed-keys]
      summary: Buat kunci sematan untuk satu situs
      operationId: create_embed_key
      x-status: implemented
      x-min-role: superadmin
      security: [{ bearerAuth: [] }]
      description: |
        Kuncinya dibuat server dan langsung aktif. Balasan membawa
        `embed_code` siap tempel.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmbedKeyCreate' }
      responses:
        '201':
          description: Kunci dibuat
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmbedKey' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/admin/embed-keys/{key}:
    parameters:
      - name: key
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [admin-embed-keys]
      summary: Ubah nama, daftar situs, atau status aktif kunci sematan
      operationId: update_embed_key
      x-status: implemented
      x-min-role: superadmin
      security: [{ bearerAuth: [] }]
      description: |
        Berlaku saat panel dibuka berikutnya. Kunci yang dinonaktifkan juga
        menolak pertanyaan dari panel yang sedang terbuka (403). Kuncinya sendiri
        tidak berubah, jadi situs penyemat tidak perlu mengganti kodenya.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmbedKeyUpdate' }
      responses:
        '200':
          description: Tersimpan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmbedKey' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }

    delete:
      tags: [admin-embed-keys]
      summary: Hapus kunci sematan secara permanen
      operationId: delete_embed_key
      x-status: implemented
      x-min-role: superadmin
      security: [{ bearerAuth: [] }]
      description: |
        Situs yang masih memasangnya harus diberi kunci baru. Percakapan dari
        situs itu tetap tersimpan, hanya tidak lagi tertaut ke kunci mana pun.
        Untuk menghentikan sementara, nonaktifkan lewat PATCH.
      responses:
        '204':
          description: Terhapus
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

# =============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token dari `POST /api/admin/login`. Hanya untuk endpoint admin.
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
    ValidationError:
      description: Body permintaan tidak sah
      content:
        application/json:
          schema: { $ref: '#/components/schemas/HTTPValidationError' }
...
    Unauthorized:
      description: |
        Token tidak ada, tidak valid, atau kedaluwarsa; akun sudah dinonaktifkan
        atau dihapus; atau kata sandi sudah diganti sejak token terbit. Frontend
        menghapus sesi dan kembali ke halaman login.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: |
        Level akun di bawah `x-min-role` operasi ini, atau staf/dosen menyentuh
        dokumen unit lain. Sesi tetap berlaku; `detail` siap tampil.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Sumber daya tidak ditemukan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    FaqEntry:
      type: object
      required: [id, question, answer, unit, updated_at, is_active, chunk_count, stale]
      properties:
        id: { type: string, format: uuid }
        question:
          type: string
          description: Sekaligus judul sumber pada kartu sitasi yang dilihat mahasiswa.
        answer: { type: string }
        unit: { type: string }
        valid_until:
          type: [string, 'null']
          format: date
          description: |
            null berarti berlaku tanpa batas. Lewat tanggal ini, entri berhenti
            terambil retrieval secara otomatis (FR-2).
        updated_at: { type: string, format: date-time }
        is_active: { type: boolean }
        chunk_count:
          type: integer
          minimum: 0
          description: |
            Jawaban pendek menjadi satu potongan; jawaban panjang dipecah, dan
            setiap potongan tetap membawa pertanyaannya.
        stale:
          type: boolean
          description: |
            Sama dengan dokumen: lebih dari 6 bulan tidak diperbarui, atau sudah
            lewat masa berlaku.
        uploaded_by:
          type: [string, 'null']
          description: Email admin yang terakhir menyimpannya.
    FaqPage:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/FaqEntry' }
        total:
          type: integer
          minimum: 0
          description: Jumlah entri yang cocok dengan filter, untuk paginasi.
    FaqEntryCreate:
      type: object
      required: [question, answer, unit]
      properties:
        question: { type: string, minLength: 5, maxLength: 500 }
        answer: { type: string, minLength: 10, maxLength: 5000 }
        unit:
          type: string
          minLength: 2
          maxLength: 200
          description: |
            Salah satu `name` dari `GET /api/units`, tanpa peduli huruf besar.
            Yang tersimpan selalu ejaan resmi; nama lain dijawab 422.
        valid_until: { type: [string, 'null'], format: date }
    FaqEntryUpdate:
      type: object
      description: Hanya field yang dikirim yang diubah.
      properties:
        question: { type: string, minLength: 5, maxLength: 500 }
        answer: { type: string, minLength: 10, maxLength: 5000 }
        unit:
          type: string
          minLength: 2
          maxLength: 200
          description: |
            Salah satu `name` dari `GET /api/units`, tanpa peduli huruf besar.
            Yang tersimpan selalu ejaan resmi; nama lain dijawab 422.
        valid_until:
          type: [string, 'null']
          format: date
          description: Kirim null untuk menjadikan entri berlaku tanpa batas.
        is_active: { type: boolean }
...
    EmbedKeyInfo:
      type: object
      required: [allowed_origins]
      properties:
        allowed_origins:
          type: array
          items: { type: string }
          description: Menjadi `frame-ancestors` halaman `/embed`. Kosong = situs mana pun.
    EmbedKey:
      type: object
      required: [key, name, allowed_origins, is_active, created_at, questions_30d]
      description: Satu kunci sematan beserta pemakaiannya.
      properties:
        key:
          type: string
          description: '`emb_` + 24 karakter. Bukan rahasia: tertulis di kode sumber situs penyemat.'
        name: { type: string, description: Nama situs, hanya untuk dashboard. }
        allowed_origins:
          type: array
          items: { type: string }
          description: Situs yang boleh memuat panel. Kosong = situs mana pun.
        is_active: { type: boolean }
        created_by: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        questions_30d:
          type: integer
          description: Pertanyaan mahasiswa dari situs ini, 30 hari terakhir.
        last_used_at:
          type: [string, 'null']
          format: date-time
          description: Pertanyaan terakhir dari situs ini; null = belum pernah dipakai.
        embed_code:
          type: [string, 'null']
          description: Baris `<script>` siap tempel; null bila `PORTAL_URL` API belum diisi.
    EmbedKeyCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        allowed_origins:
          type: array
          maxItems: 20
          items: { type: string }
          description: |
            Alamat situs lengkap tanpa path, mis. `https://pmb.instiki.ac.id`.
            Wildcard subdomain `https://*.instiki.ac.id` diterima. Garis miring
            akhir, huruf besar, dan port bawaan dirapikan; duplikat dibuang.
            Kosong = situs mana pun.
    EmbedKeyUpdate:
      type: object
      description: Hanya field yang dikirim yang diubah. Field tak dikenal ditolak.
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        allowed_origins:
          type: array
          maxItems: 20
          items: { type: string }
        is_active: { type: boolean }
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: |
            Kalimat siap tampil. Untuk endpoint admin bahasanya non-teknis
            (PRD §9: "dokumen sedang diproses", bukan "embedding in progress").
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            type: [string, integer]
        msg: { type: string }
        type: { type: string }
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items: { $ref: '#/components/schemas/ValidationError' }